backend/requirements.txt
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dlib==19.24.2
filelock==3.13.1
Flask==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
numpy==1.26.3
opencv-contrib-python==4.9.0.80
platformdirs==4.1.0
polygon-api-client==1.13.4
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==1.26.18
virtualenv==20.25.0
websockets==12.0
Werkzeug==3.0.1
